+++ server/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Pandas: the `pivot` step now keeps null values in index columns and handle them as any other values.
+
 ## [0.48.6] - 2024-12-11
 
 ### Fixed